refactor(budget): extract budget-value and budget-defaults modules
- Extract pure utility functions into budget-value.ts
- Extract default allocation logic into budget-defaults.ts
- Narrow resolveDefaultValueForSlug param type (remove unsafe cast)
- Wrap applyDefaultAllocation upserts in a transaction
- Guard --defaults from being combined with positional args
- Fix (c) abbreviation to (category) in filter
- Update default necessities from 30% to 50% (totals 100%)
